CR UPDATES: removed shuf_start - shuf driver now all in shuf.c, test_shuf now using fork/exec to run tests rather than function calls, refactoring in shuf.c, updated Makefile to reflect changes
