+++ Makefile
@@ -87,7 +87,7 @@ $U/initcode: $U/initcode.S
 tags: $(OBJS) _init
 	etags *.S *.c
 
-ULIB = $U/ulib.o $U/usys.o $U/printf.o $U/umalloc.o $U/vec.o $U/shuffler.o $U/shuf_start.o
+ULIB = $U/ulib.o $U/usys.o $U/printf.o $U/umalloc.o $U/vec.o $U/shuffler.o
 
 _%: %.o $(ULIB)
 	$(LD) $(LDFLAGS) -T $U/user.ld -o $@ $^
